+++ RELEASE_NOTES.md
@@ -1,7 +1,7 @@
 Version 1.7.0
 -------------------
 
-- added `pretty_print()` functions several classes: `DataSource`, `DataSourceStatistics`, `ConjunctionSearch`
+- added `pretty_print()` functions several classes: `DataSource`, `DataSourceStatistics`, `ConjunctionSearch`, `DataProductSearch`
 
 
 Version 1.6.0